+++ src/main/java/com/coolerpromc/bettercampfirepot/block/entity/BetterCampfireBlockEntity.java
@@ -14,7 +14,9 @@ import com.coolerpromc.bettercampfirepot.block.BetterCampfireBlock;
 import com.coolerpromc.bettercampfirepot.item.BetterCampfirePotItem;
 import com.coolerpromc.bettercampfirepot.menu.CookingPotMenu;
 import com.coolerpromc.bettercampfirepot.recipe.BetterCampfirePotRecipe;
+import com.coolerpromc.bettercampfirepot.util.CampfirePotSlot;
 import com.coolerpromc.bettercampfirepot.util.KotlinHelper;
+import com.coolerpromc.bettercampfirepot.util.Side;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.datafixers.util.Pair;
 import net.minecraft.client.multiplayer.ClientLevel;
@@ -31,6 +33,7 @@ import net.minecraft.resources.RegistryOps;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundSource;
+import net.minecraft.util.ExtraCodecs;
 import net.minecraft.util.FastColor;
 import net.minecraft.world.MenuProvider;
 import net.minecraft.world.SimpleContainer;
@@ -44,6 +47,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.HorizontalDirectionalBlock;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.gameevent.GameEvent;
@@ -81,6 +85,7 @@ public class BetterCampfireBlockEntity extends BlockEntity implements MenuProvid
     public int time = 0;
     public BetterCampfirePotRecipe currentRecipe;
     public int progressPerTick = 2;
+    private Map<Side, CampfirePotSlot> capabilityBySide = new HashMap<>();
 
     public List<ItemStack> lastSeasoningStacks = new ArrayList<>();
 
@@ -156,6 +161,16 @@ public class BetterCampfireBlockEntity extends BlockEntity implements MenuProvid
 
     public BetterCampfireBlockEntity(BlockPos pos, BlockState state) {
         super(BetterCampfirePot.BETTER_CAMPFIRE_BE.get(), pos, state);
+        initCapabilityBySide();
+    }
+
+    private void initCapabilityBySide(){
+        capabilityBySide.putIfAbsent(Side.TOP, CampfirePotSlot.SEASONING);
+        capabilityBySide.putIfAbsent(Side.BOTTOM, CampfirePotSlot.OUTPUT);
+        capabilityBySide.putIfAbsent(Side.FRONT, CampfirePotSlot.INPUT);
+        capabilityBySide.putIfAbsent(Side.BACK, CampfirePotSlot.INPUT);
+        capabilityBySide.putIfAbsent(Side.LEFT, CampfirePotSlot.INPUT);
+        capabilityBySide.putIfAbsent(Side.RIGHT, CampfirePotSlot.INPUT);
     }
 
     @OnlyIn(Dist.CLIENT)
@@ -443,6 +458,10 @@ public class BetterCampfireBlockEntity extends BlockEntity implements MenuProvid
         tag.put("seasoningHandler", seasoningHandler.serializeNBT(registries));
         tag.put("outputHandler", outputHandler.serializeNBT(registries));
         tag.putInt("progressPerTick", progressPerTick);
+        if (capabilityBySide.isEmpty()){
+            this.initCapabilityBySide();
+        }
+        tag.put("capabilityBySide", ExtraCodecs.strictUnboundedMap(Side.CODEC, CampfirePotSlot.CODEC).encodeStart(NbtOps.INSTANCE, capabilityBySide).getOrThrow());
 
         if (potComponent != null) {
             PotComponent.Companion.getCODEC().encodeStart(NbtOps.INSTANCE, potComponent).result().ifPresent(encoded -> tag.put("PotComponent", encoded));
@@ -461,6 +480,10 @@ public class BetterCampfireBlockEntity extends BlockEntity implements MenuProvid
         seasoningHandler.deserializeNBT(registries, tag.getCompound("seasoningHandler"));
         outputHandler.deserializeNBT(registries, tag.getCompound("outputHandler"));
         this.progressPerTick = tag.getInt("progressPerTick");
+        this.capabilityBySide = new HashMap<>(ExtraCodecs.strictUnboundedMap(Side.CODEC, CampfirePotSlot.CODEC).parse(NbtOps.INSTANCE, tag.getCompound("capabilityBySide")).getOrThrow());
+        if(capabilityBySide.isEmpty()){
+            this.initCapabilityBySide();
+        }
 
         if (tag.contains("PotComponent")) {
             PotComponent.Companion.getCODEC().parse(NbtOps.INSTANCE, tag.getCompound("PotComponent")).result().ifPresent(component -> this.potComponent = component);
@@ -545,14 +568,33 @@ public class BetterCampfireBlockEntity extends BlockEntity implements MenuProvid
     }
 
     public @Nullable IItemHandler getCapability(@Nullable Direction direction) {
-        if (direction == Direction.DOWN){
-            return outputHandler;
-        }
-        else if (direction == Direction.UP){
-            return seasoningHandler;
-        } else if (direction != null) {
-            return inputHandler;
-        }
-        return null;
+        if (direction == null) return null;
+
+        Direction facing = getBlockState().getValue(HorizontalDirectionalBlock.FACING);
+        Side side = Side.fromDirection(direction, facing);
+
+        CampfirePotSlot slot = getCapabilityBySide(side);
+        if (slot == null) return null;
+
+        return getItemHandlerForSlot(slot);
+    }
+
+    private IItemHandler getItemHandlerForSlot(CampfirePotSlot slot){
+        return switch (slot){
+            case INPUT -> inputHandler;
+            case SEASONING -> seasoningHandler;
+            case OUTPUT -> outputHandler;
+        };
+    }
+
+    public CampfirePotSlot getCapabilityBySide(Side side){
+        return capabilityBySide.get(side);
+    }
+
+    public void setCapabilityBySide(Side side, CampfirePotSlot slot) {
+        this.capabilityBySide.put(side, slot);
+        setChanged();
+        level.sendBlockUpdated(getBlockPos(), getBlockState(), getBlockState(), 3);
+        level.invalidateCapabilities(getBlockPos());
     }
 }